* Use double precision for position computation due to round problems  * iPhone 3G GPS provide speed in firmware 2.2. xGPS used on firmware 2.2 will now use this functionality
git-svn-id: https://svn.xwaves.net/xGPS/software/trunk/xGPS@100 073a2a62-4f4e-0410-ac76-ce2687731022

+++ Classes/DirectionsController.h
@@ -24,7 +24,7 @@
 @protocol DirectionsControllerDelegate
 -(void)directionsGot:(NSString*)from to:(NSString*)to error:(NSError*)err;
 -(void)nextDirectionChanged:(Instruction*)instr;
--(void)nextDirectionDistanceChanged:(float)dist;
+-(void)nextDirectionDistanceChanged:(double)dist;
 @end
 
 @interface DirectionsController: NSObject<DrivingInstructionMovingProtocol> {
